moved librsvg [extra]->[testing]
Unbeknownst to me there were a massive slew of un-git'd packages
in [testing] for a proposed gtk2 upgrade.  Some of these packages
are out of date, so when I built new versions I didn't get any
warnings, but apparently I silently put newer versions in
[extra] than there were in [testing].  Now I've gone through
and just replaced the [testing] copies, so everything should be fine.